remove DataInterface.ts from common
