+++ client/src/routes/_auth/room_.$roomId.lazy.tsx
@@ -7,10 +7,11 @@ import {
 } from '@/components/chatUtilities/Messages';
 import { InputBar } from '@/components/InputBar';
 import { useQuasmStore } from '@/lib/quasmStore';
+import { MessageDetails, StoryChunkDetails } from '@/lib/sharedTypes';
 
 function Story() {
-  const messages = useQuasmStore(state => state.messages);
-  const story = useQuasmStore(state => state.story);
+  const messages = useQuasmStore(state => state.messages as MessageDetails[]);
+  const story = useQuasmStore(state => state.story as StoryChunkDetails[]); // temporary solution but it has to be changed to sockets anyway
   return (
     <div className='flex h-full flex-col justify-end'>
       <OutletWrapper>
